Hugo's config.toml renamed to hugo.toml.
From Hugo's documentation (see
https://gohugo.io/getting-started/configuration/#configuration-file):

With v0.109.0 and earlier the basename of the site configuration file
was config instead of hugo. You can use either, but should transition to
the new naming convention when practical.

As we don't have control on the hugo version used by all the learning
paths developpers, we renamed config.toml to hugo.toml in order to be
future proof (this is the new default name), and updated all our scripts
and documentation to explicitely pass the configuration file name with
the '--config hugo.toml' command line option.